+++ src/render/utilities/color_math2.c
@@ -11,12 +11,12 @@ t_color	clr_mult(t_color color, double val)
 	return (create_color(color.r * val, color.g * val, color.b * val));
 }
 
-t_color	clr_sub(t_color color, int val)
+t_color	clr_sub(t_color color, double val)
 {
 	return (create_color(color.r - val, color.g - val, color.b - val));
 }
 
-t_color	clr_add(t_color color, int val)
+t_color	clr_add(t_color color, double val)
 {
 	return (create_color(color.r + val, color.g + val, color.b + val));
 }
